feat(wave-23/J.beta): restore upstream stat widths (int32 across MoveEffect, SpawnPowerEntry, PowerInstance, CardCounts, CompactState.round) + Stat::pack16 + widened Zobrist key tables
Per Q2-ADR-014 (wave-23): upstream STS2 (Godot/C#) uses uniform int (32-bit
signed) for all combat stats. Q2's narrower types (int16_t for damage and
power stacks; uint8_t for HP via pack8, card counts; uint16_t for round)
diverged from upstream contract.

Widened:
- MoveEffect.value, SpawnPowerEntry.stacks, PowerInstance.stacks: int16 -> int32
- MonsterMoveTable.min_hp/max_hp: uint8 -> int32
- CardCounts.counts[]: uint8 -> int32
- CompactState.round_: uint16 -> int32
- Stat::pack8() -> pack16() (asserts v in [0, 65535]; uint16 return)
- Reordered struct fields for natural alignment: MoveEffect 8B, SpawnPowerEntry 8B, PowerInstance 8B; CardCounts 20B (5 * int32).

Zobrist key tables widened: kMaxHp 256->1024, kMaxBlock 256->1024,
kMaxStacks 100->256, kMaxCountPerCardZone 16->64. ~4 MB additional static
table memory; TT entries unaffected (38 B/entry post-J.alpha LRU revert).

pack_counts: DELETED. Audit found no consumers in src/ or tests/; the prior static_assert was a guard for a hypothetical uint64 packing helper that never materialized. Cleanly removed alongside CardCounts widening.

Narrow-arithmetic audit: ~65 hits across src + tests + include; 19 STAT-ARITHMETIC widened (powers helpers, builder casts, projection casts, test value brace-init), 31 INDEX/PROTOCOL/ENUM-CAST preserved (uint8_t flag bitmasks, MoveId/MonsterKind/Phase enum->uint8_t indexing, find_move_index 0xFF sentinel, follow_up_index/branch_weights protocol fields).

Cultist Zobrist BYTE ROTATED per Q2-ADR-010 §Recovery. NEW constants:
  kCultistZobristKeyLo = 0x569115efa81a95dc
  kCultistZobristKeyHi = 0x9a06f1e505846a80
Cultist + Louse pin VALUES BIT-IDENTICAL (expected_hp=40.9083 /
expected_rounds=6.45798; expected_hp=0.0407931 / expected_rounds=10.152).

algorithm_sha source list extended: stat.h added per Q2-ADR-005 amendment
(docs consolidated in J.gamma).

Q2-ADR-013 Amendment 5 (LRU revert; J.alpha) + Q2-ADR-014 (this) ratified
in J.gamma docs stream.

+++ engine/cpp/src/ai/probability.cc
@@ -46,7 +46,7 @@ void enumerate_recursive(const CardCounts& pool, int k_left,
   const int avail = pool[id];
   const int upper = std::min(avail, k_left);
   for (int take = 0; take <= upper; ++take) {
-    acc[id] = static_cast<uint8_t>(take);
+    acc[id] = take;
     enumerate_recursive(pool, k_left - take, card_idx + 1,
                         numerator * binom(avail, take), acc, out, inv_denom);
   }
